+++ asciidoctor-mathx-extension/src/main/groovy/org/asciidoctor/mathx/BlockProcessor.groovy
@@ -16,13 +16,12 @@
 package org.asciidoctor.mathx
 
 import groovy.transform.CompileStatic
-import org.asciidoctor.ast.AbstractBlock
-import org.asciidoctor.ast.DocumentRuby
+import org.asciidoctor.ast.StructuralNode
 import org.asciidoctor.extension.BlockProcessor as AsciidoctorBlockProcessor
 import org.asciidoctor.extension.Reader
+
 import javax.imageio.ImageIO
 import java.awt.image.BufferedImage
-
 /**
  * This processor is going to transform a listing paragraph block into
  * an image with a mathematical expression
@@ -32,103 +31,108 @@ import java.awt.image.BufferedImage
 @CompileStatic
 class BlockProcessor extends AsciidoctorBlockProcessor {
 
+    static final String TEMPLATE_TITLE = 'title="$it"'
+    static final String TEMPLATE_ALT = 'alt="$it"'
+    static final String TEMPLATE_ALIGN ='align="$it"'
+    static final String TEMPLATE_WIDTH ='width=$it'
+    static final String TEMPLATE_HEIGHT = 'height=$it'
+
+    static final Closure<Object> ONLY_WITH_VALUE = { Map.Entry<String, Object> entry -> entry.value }
+
+    static final String DOCUMENT_ATTR_IMAGES_DIR = 'imagesdir'
+    static final String DOCUMENT_ATTR_GRADLE_PROJECT_DIR = 'gradle-projectdir'
+
+    static final String PROCESSOR_NAME = 'mathx'
+    static final Map<String,Object> PROCESSOR_CONFIG = [contexts: [':listing', ':paragraph']] as Map<String, Object>
+
     /**
-     * Establishes the pattern under which this processor is going to be called passing the required
-     * values to the super constructor. Every time there's a block named 'mathx' and of type
-     * `listing` + `paragraph` this processor will be hit.
-     *
-     * @since 0.1.0
-     */
+      * Establishes the pattern under which this processor is going to be called passing the required
+      * values to the super constructor. Every time there's a block named 'mathx' and of type
+      * `listing` + `paragraph` this processor will be hit.
+      *
+      * @since 0.1.0
+      */
     BlockProcessor() {
-        super('mathx', [contexts: [':listing', ':paragraph']] as Map<String, Object>)
+        super(PROCESSOR_NAME, PROCESSOR_CONFIG)
     }
 
     @Override
-    Object process(AbstractBlock parent, Reader reader, Map<String, Object> blockAttributes) {
-        DocumentRuby document = parent.document
-        Map<String, Object> documentAttributes = document.attributes
+    Object process(StructuralNode node, Reader reader, Map<String, Object> blockAttrs) {
+        Integer width = blockAttrs.width as Integer
+        Integer height = blockAttrs.height as Integer
+        String align = blockAttrs.align
+        String title = blockAttrs.title
+        String name = blockAttrs.name
 
         String formula = reader.readLines().join('\n')
-        String hash = Utils.getMD5(formula)
-        File realImageFile = getRealImageFile(hash, documentAttributes)
+        File file = createImagePath(node, formula, name)
+        BufferedImage image = createImage(file, formula, width, height)
 
-        if (!realImageFile.exists()) {
-            BufferedImage image = Utils.TryOrLogError("Error while generating mathx image") {
-                Utils.createImageFromLatex(formula, blockAttributes)
-            }
-
-            ImageIO.write(image, "png", realImageFile)
+        if (!image) {
+            return null
         }
 
         Map<String,Object> imageAttrs = [
-            target: realImageFile.name,
-            alt: realImageFile.name,
-            title: blockAttributes.title,
-            width: blockAttributes.width,
-            height: blockAttributes.height
-        ]
+            target: file.name,
+            alt: file.name,
+            title: title,
+            align: align,
+            width: width,
+            height: height
+        ] as Map<String, Object>
 
-        return createBlock(parent, "image", "", imageAttrs, Constants.EMPTY_MAP_OBJ_OBJ)
+        List<String> lines = createImageLines(imageAttrs.findAll(ONLY_WITH_VALUE))
+        parseContent(node, lines)
+
+        return null
     }
 
-    /**
-     * Resolves where to write the latex image
-     *
-     * @param formulaHash hash which be used as the file name
-     * @param documentAttributes {@link DocumentRuby} attributes*
-     * @return the real path of the image file
-     */
-    private File getRealImageFile(String formulaHash, Map<String, Object> documentAttributes) {
-        String documentDir = documentAttributes.docdir?.toString()
-        String imagesOutputDir = documentAttributes.imagesoutdir?.toString()
-        String imagesDir = documentAttributes.imagesdir?.toString()
+    private static BufferedImage createImage(File imageFile, String formula, Integer width, Integer height) {
+        return Utils.TryOrLogError("Error while generating mathematical expression image") {
+            BufferedImage image = Utils.createImageFromLatex(formula)
+            BufferedImage resized = Utils.resizeImage(image, width, height)
 
-        File images = imagesOutputDir
-          ? getOutputImagesDir(documentDir, imagesOutputDir)
-          : getImagesDir(imagesDir)
+            ImageIO.write(resized, "png", imageFile)
 
-        String filename= "${formulaHash}.png"
-        File realImagePath = new File(images, filename)
-
-        return realImagePath
-    }
-
-    /**
-     * Resolves the real images dir from the 'images' entry in document
-     * configuration
-     *
-     * @param imagesPath image path found in document configuration
-     * @return where to find the real
-     */
-    private File getImagesDir(String imagesPath) {
-        File imagesDir = new File(imagesPath)
-
-        if (imagesDir.isAbsolute()) {
-            return imagesDir
+            return resized
         }
-
-        return new File(this.rubyRuntime.currentDirectory, imagesPath)
     }
 
-    /**
-     * Resolves the real images dir from the 'imagesoutput' entry in document
-     * configuration
-     *
-     * Resolves where is the real image file given an output dir. If the
-     * output directory is absolute, then it's used, otherwise we resolve
-     * where the document is and then we concatenate the output dir to it
-     *
-     * @param documentDir where is the document
-     * @param imagesOutputPath the declared output dir in configuration
-     * @return the output dir where to find the image
-     */
-    private static File getOutputImagesDir(String documentDir, String imagesOutputPath) {
-        File imagesOutputDir = new File(imagesOutputPath)
+    private static List<String> createImageLines(Map<String, Object> attrs) {
+        String target = attrs.target
+        String caption = attrs.title ? ".${attrs.title}" : null
 
-        if (imagesOutputDir.isAbsolute()) {
-            return imagesOutputDir
-        }
+        String title = Utils.renderIfPresent(attrs.title, TEMPLATE_TITLE)
+        String alt =  Utils.renderIfPresent(attrs.alt, TEMPLATE_ALT)
+        String align = Utils.renderIfPresent(attrs.align, TEMPLATE_ALIGN)
+        String width = Utils.renderIfPresent(attrs.width, TEMPLATE_WIDTH)
+        String height = Utils.renderIfPresent(attrs.height, TEMPLATE_HEIGHT)
 
-        return new File(documentDir, imagesOutputPath)
+        String properties = [alt, width, height, title, align]
+                .grep()
+                .join(',')
+                .toString()
+
+        // In order to get an auto-numbered title with caption (e.g., Figure N. Title), and make the alignment to work
+        // you need to use the block image macro.
+        // The block image macro (like all block macros) uses two colons instead of one.
+        return [caption, "image::$target[$properties]"].grep()*.toString()
+    }
+
+    private static File createImagePath(StructuralNode node, String formula, String customName) {
+        String filename = "${Utils.getMD5(formula)}.png"
+        String finalName = customName ?: filename
+
+        Map<String,Object> attrs = node.document.attributes
+        String projectDir = attrs[DOCUMENT_ATTR_GRADLE_PROJECT_DIR]
+        String imagesDir = attrs[DOCUMENT_ATTR_IMAGES_DIR]
+
+        String filePath = new File("$projectDir/$imagesDir", finalName).absolutePath
+        String file = node.normalizeWebPath(filePath, "", true)
+        File fileToCreate = new File(file)
+
+        fileToCreate.parentFile.mkdirs()
+
+        return fileToCreate
     }
 }